backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pyjwt[crypto]>=2.10.1",
    "pymongo>=4.14.1",
    "pydantic-settings>=2.10.1",
    "argon2-cffi>=25.1.0",
]

[dependency-groups]
dev = [
    "coverage>=7.10.7",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "testcontainers>=4.13.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
